PostMenu, CommentMenu을 하나의 Menu 컴포넌트로 통합 (#708)
* refactor: (#687) 댓글 및 게시글 action 관련 상수 및 타입을 constants 및 types 폴더로 분리

* chore: (#687) 모듈 오타 수정

* refactor: (#687) Menu 컴포넌트 구현 및 기존의 CommentMenu, PostMenu 컴포넌트 삭제

* refactor: (#687) 컴포넌트 이름을 Menu에서 PostMenu 로 수정, 스토리 이름 수정

* refactor: (#687) 타입명을 컴포넌트명과 통일

* refactor: (#687) PostMenu의 props type을 제네릭 T로 대체
타입의 범용성 개선, 컴포넌트의 범용성 개선
