+++ frontend/src/components/common/Menu/types.ts
@@ -1,0 +1,5 @@
+export interface MenuItem<T> {
+  content: string;
+  color: 'black' | 'red';
+  action: T;
+}
